% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_precomputations.R
\name{apply_precomputations}
\alias{apply_precomputations}
\title{Apply precomputations to data}
\usage{
apply_precomputations(data, precomputations_list)
}
\arguments{
\item{data}{A \code{data.frame} containing the preprocessed patient survey data.}

\item{precomputations_list}{A \code{list} of named precomputations. Each component is also a \code{list}
containing a computation rule \code{.$expr} and value labels \code{.$label}.}
}
\value{
This takes the patient survey data and performs the specified precomputations by adding new columns equal
to the number of precomputations to the data set. Each new column is named after one precomputation and filled with values
as obtained from the corresponding expression \code{.$expr} applied (rowwise) to the data set.
The extended survey data set is returned as a \code{data.frame}.
}
\description{
Apply a specified set of precomputation rules to preprocessed patient survey data
}
