% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_results.R
\name{compute_results}
\alias{compute_results}
\title{Compute all QIs in a QIDB data pair.}
\usage{
compute_results(
  qidb_data_pair,
  conf_level = 0.95,
  a = 0.5,
  b = 0.5,
  nMC = 1e+05
)
}
\arguments{
\item{qidb_data_pair}{A list of QIs with linked data (of class \code{qi})}

\item{conf_level}{Confidence level of the resulting uncertainty interval}

\item{a}{first prior parameter of Beta distribution for each QI Merkmal}

\item{b}{second prior parameter of Beta distribution for each QI Merkmal}

\item{nMC}{Number of Monte Carlo Samples (needed for QIs with >1 Merkmale)}
}
\value{
A \code{data.frame} containing all computation results, one row per result
}
\description{
Applies the Bayesian computation method to all QIs in a QIDB-data-pair, separately to all LEs as
identified through the \code{meta_unit} column in the data.
}
\section{Fields}{

\describe{
\item{\code{KN_ID}}{the idenifier of the QI}

\item{\code{meta_unit}}{the identifier of the Leistungserbringer (LE)}

\item{\code{QI_hat}}{Posterior mean estimate for QI value}

\item{\code{J}}{overall number of patients in the data set.}

\item{\code{lower}}{lower border of the uncertainty interval.}

\item{\code{upper}}{upper border of the uncertainty interval.}

\item{\code{Auff_stat}}{indicates if result is "statistisch auffaellig"}

\item{\code{RefVal}}{reference value for computation of "statistisch auffaellig"}
}}

