% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qi.R
\name{connect_qidb2data}
\alias{connect_qidb2data}
\title{Construct a QIDB data pair}
\usage{
connect_qidb2data(qidb, data)
}
\arguments{
\item{qidb}{A \code{list} containing specification of multiple QIs. Each list component represents the raw information to one QI.}

\item{data}{A \code{data.frame} containing the pre-processed patient survey data the QI should be evaluated on.}
}
\value{
The patient survey data is joined to every qi from the provided \code{qidb}, resulting in a \code{list} of objects of class \code{qi}.
}
\description{
Generate a full set of computable QIs joined by corresponding data.
}
