% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qi.R
\name{join_data2qi}
\alias{join_data2qi}
\title{Make a QI data pair}
\usage{
join_data2qi(qi, data)
}
\arguments{
\item{qi}{a list containing necessary QI specifications}

\item{data}{a data.frame containing the pre-processed patient survey data the QI should be evaluated on}
}
\value{
The function takes the QI specifications and data and generates a more extensive list object of class \code{qi} to be used for computation of the QI result.
For instance, the output contains some meta-information such as number of Merkmale or item categories. The output is ready to use for QI evaluation, i.e.
can serve as input for the function \code{evaluateQI}.
}
\description{
Generate a pair of QI specifications and corresponding data and compute some meta information.
}
