% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mappings_pci}
\alias{mappings_pci}
\title{PCI response category mappings}
\format{
A \code{list} containing two objects. The object \code{point_mappings}, which is itself a list containing
four unnamed list components specifying remapping rules for the answer categories.
Each of these four lists contains two objects:
\describe{
  \item{mapping}{A named \code{numeric} vector. The categories provided by the names are mapped onto the corresponding values between 0 and 100.}
  \item{felder}{A \code{character} vector providing the column names, which the \code{mapping} should be applied on.}
}
The second object \code{ausweichkategorien} is also a list containing specifications of non-informative values given through five lists of the form:
\describe{
  \item{value}{A \code{numeric} value, which should be mapped to NA}
  \item{felder}{A \code{character} vector providing the column names, for which the NA replacement should be applied.}
}
}
\usage{
mappings_pci
}
\description{
A list of item category mapping rules applied to patient survey data for the quality assurance domain PCI (percutaneous coronary intervention and coronary angiography).
}
\keyword{datasets}
