% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precomputations_pci}
\alias{precomputations_pci}
\title{PCI precomputations}
\format{
A \code{list} containing two named precomputations. The component names will be used as name of the new data column
that will be generated by the precomputation. Each precomputation is itself a list containing 3 objects:
\describe{
  \item{expr}{The computation rule expression. This commonly refers to specific columns in the underlying data set and utilizes operators from the \code{iqtigfunctions} package.}
  \item{prototype}{An optional reference column. Value labels for the newly computed field are inherited from this reference column.}
  \item{labels}{(optional) A set of labels for the newly computed column values}
}
}
\usage{
precomputations_pci
}
\description{
A list of precomputations applied to patient survey data for the quality assurance domain PCI (percutaneous coronary intervention and coronary angiography).
Each precomputation generates a new column within the data.
}
\keyword{datasets}
