% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qi.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data for the QI-computation.}
\usage{
prepareData(qi, clean_data)
}
\arguments{
\item{qi}{A qi (i.e. class \code{qi})}

\item{clean_data}{A \code{data.frame} containing the cleaned patient survey data
(all non-missing values are relevant for computation)}
}
\value{
Takes the cleaned data set, applies GG condition, restricts to relevant
columns and returns the resulting data set as a \code{data.frame}.
}
\description{
Data preparation includes restricting to the GG, defining the static
part of the weights, etc. We introduce this data preparation step
in order to be able to do subsequent bootstrap resampling faster.
}
