% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Apply category mappings}
\usage{
preprocess_data(raw_data, mappings = NULL)
}
\arguments{
\item{raw_data}{A \code{data.frame} containing the raw patient survey data.}

\item{mappings}{A \code{list} containing the mapping specifications, i.e. points value mappings and affected data fields.}
}
\value{
Takes the survey data \code{dat} and applies the point mappings as listed in \code{mappings$point_mappings} and
the replacement mappings in \code{mappings$ausweichkategorien}.
Additional \code{ID} and \code{meta_unit} columns are also added to the output.
}
\description{
Create a new data set from raw survey data by applying specified mappings of certain variables
}
