% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qidb_pci}
\alias{qidb_pci}
\title{PCI quality indicator data base}
\format{
A \code{list} containing a set of quality indicators. Each quality indicator is itself list containing 7 objects:
\describe{
  \item{Name}{The indicator name}
  \item{KN_ID}{The indicator ID}
  \item{GG}{The patient population ('Grundgesamtheit') to be included in the quality indicator, provided as a filter expression.}
  \item{RefArt}{The type of reference value. Possible values are "Fest" (fixed) or "Kein Referenzwert" (no reference value).}
  \item{RefVal}{The reference value to be used for indicator result classification}
  \item{RefOp}{The reference operator indicating, wether results greater (">=") or lower ("<=") than the reference value are desirable.}
  \item{Merkmale}{A \code{list} of attributes, that constitute the quality indicator. Each attribute is a list containing the attribute's \code{Name} and the underlying \code{Items}.}
}
}
\usage{
qidb_pci
}
\description{
A list of quality indicator specifications for the quality assurance domain PCI (percutaneous coronary intervention and coronary angiography).
}
\keyword{datasets}
