% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_data_pci}
\alias{raw_data_pci}
\title{PCI patient survey data}
\format{
A \code{data.frame} with 109 rows and 21 variables, which include 14 survey items:
\describe{
  \item{ID}{ID of data entry}
  \item{ID_LE}{ID of the patient's care provider}
  \item{Fragebogen}{Questionnaire type}
  \item{FB_Sent}{Date of questionnaire submission}
  \item{Gebdatum}{Date of birth}
  \item{Geschlecht}{Sex}
  \item{BMI}{Body mass index}
  \item{ARermutigtn}{Survey item: "Doctors encouraged me to ask questions during a consultation."}
  \item{ARernstn}{Survey item: "My concerns were taken seriously."}
  \item{ARrespektn}{Survey item: "Doctors treated me with respect."}
  \item{ARgelegenheitn}{Survey item: "I was given the opportunity to talk to a doctor, if I had any question."}
  \item{ARInfverstn}{Survey item: "The information I was given by doctors were comprehensible.}
  \item{ARfachwortn}{Survey item: "When talking to me doctors used medical terms that I did not comprehend."}
  \item{ARdeutschn}{Survey item: "Sometimes I cound not communicate with doctors, because they did not speak german."}
  \item{ARgesprochenn}{Survey item: "In my presence doctors talked about me as if I was not there."}
  \item{PAvoranginan}{Survey item: "Did you have any angina pectoris related problems or shortness of breath in advance of the procedure?"}
  \item{Anginaruhen}{Survey item: "Angina pectoris related problems: while resting."}
  \item{Anginaleichtn}{Survey item: "Angina pectoris related problems: during light everyday activities."}
  \item{PAvorbeeintrn}{Survey item: "How much was your everyday life affected by angina pectoris related problems or shortness of breath in advance of the procedure?"}
  \item{Anginaschwern}{Survey item: "Angina pectoris related problems: during heavy everyday activities."}
  \item{Anginaaussergn}{Survey item: "Angina pectoris related problems: during extraordinary physical efforts."}
}
}
\usage{
raw_data_pci
}
\description{
An artificial dataset containing patient characteristics and raw survey answers for the quality assurance domain PCI (percutaneous coronary intervention and coronary angiography).
Each data entry refers to one patient. Patient responses to survey items can take prespecified values between 0 and 100, where 0 generally refers to
"complete disagreement" or "not at all" and 100 refers to "full agreement" or "completely", depending on the item. All survey items
allow one or more special response categories for cases, in which the patient cannot or does not want to answer the survey question. These special
categories are "-99: no statement", "-98: I do not know or cannot remember", "-97: does not apply due to erroneous questionnaire filter" and "-96: question does not apply for me",
"-90: I could not do that, because of other reasons".
}
\keyword{datasets}
